LayoutTests/fullscreen/viewer-restore-scroll-position-expected.txt

This tests that page scroll is restored after viewer mode.


EVENT(load)
EVENT(fullscreenchange)
EVENT(fullscreenchange)
EXPECTED ((document.scrollingElement.scrollTop === originalScroll) == 'true') OK
